refactor: 각 도메인에 컨트롤러, 서비스, 리포지토리 폴더 추가
