--- backend/cafe-api/src/main/java/com/guruja/cafe_api/order/AdminController.java
@@ -1,4 +1,0 @@
-package com.guruja.cafe_api.order;
-
-public class AdminController {
-}
