+++ backend/cafe-api/src/main/java/com/guruja/cafe_api/order/controller/AdminController.java
@@ -1,0 +1,4 @@
+package com.guruja.cafe_api.order.controller;
+
+public class AdminController {
+}
