+++ backend/cafe-api/src/main/java/com/guruja/cafe_api/order/controller/OrderController.java
@@ -1,4 +1,4 @@
-package com.guruja.cafe_api.order;
+package com.guruja.cafe_api.order.controller;
 
 import org.springframework.web.bind.annotation.RestController;
