--- backend/cafe-api/src/main/java/com/guruja/cafe_api/product/ProductRepository.java
@@ -1,4 +1,0 @@
-package com.guruja.cafe_api.product;
-
-public interface ProductRepository {
-}
