+++ backend/cafe-api/src/main/java/com/guruja/cafe_api/product/controller/ProductController.java
@@ -1,0 +1,4 @@
+package com.guruja.cafe_api.product.controller;
+
+public class ProductController {
+}
