+++ backend/cafe-api/src/main/java/com/guruja/cafe_api/product/repository/ProductRepository.java
@@ -1,0 +1,4 @@
+package com.guruja.cafe_api.product.repository;
+
+public interface ProductRepository {
+}
