+++ backend/cafe-api/src/main/java/com/guruja/cafe_api/product/service/ProductService.java
@@ -1,0 +1,4 @@
+package com.guruja.cafe_api.product.service;
+
+public class ProductService {
+}
